update to 1.4.0 (full refactoring)

+++ InAppStorySdk/inappstorysdk/src/main/java/com/inappstory/sdk/stories/api/models/dialogstructure/BorderStructure.java
@@ -1,7 +1,7 @@
 package com.inappstory.sdk.stories.api.models.dialogstructure;
 
 public class BorderStructure {
-    public int radius;
-    public int width;
+    public float radius;
+    public float width;
     public String color;
 }
